Add files via upload
A long transmission line can be considered to have a high amount of capacitance and inductance distributed across the entire length of the line. When current drawn from the distributed capacitance of the line is greater than the current associated with the load at the receiving end, Ferranti effect occurs. Therefore in this experiment, Ferranti effect for different length of transmission lines have been determined.